bug fixes and added recording
Fixed a bug in hard_wall_case_green in the computation of g_tilde(n,m,t), there was a sign error in the case n>m.

Added a recording for the hard wall case using Green's function.

some cosmetic changes, removed some useless code that was commented out, added a few comments to the code.